fix(db): widen active_listing.id to bigint; let sale_history ids wrap
active_listing.id exhausted its int4 sequence on 2026-09-12 and froze the
listing board; #1463 bought time in the negative range (~2028-01). Widen the
column (and the dead materia_listing FK column) to bigint and move the
sequence to 2^31, above every id int4 could have issued, so live pre-outage
rows can never collide. Every i32 that carried a listing id follows, and
ClickHouse listing_events.pg_listing_id is retyped to Int64 in place with a
non-blocking mutation.

sale_history.id is the same int4 serial on a ~1.9B-row append-only table, so
its headroom is at most a few hundred million ids. A bigint rewrite there is
hours of lock on ~272 GB and ClickHouse refuses to retype pg_id in the sales
sorting keys, so instead let the sequence CYCLE into the unused negative
half: remaining positive ids are still used, the ceiling wraps instead of
failing, and the next limit (~2.1B sales out) is documented.

+++ ultros-api-types/src/lib.rs
@@ -140,7 +140,7 @@ mod tests {
     use crate::websocket::{EventType, ListingEventData, SaleEventData};
     use chrono::{DateTime, NaiveDateTime};
 
-    fn test_listing(id: i32, item_id: i32, price: i32) -> (ActiveListing, Retainer) {
+    fn test_listing(id: i64, item_id: i32, price: i32) -> (ActiveListing, Retainer) {
         (
             ActiveListing {
                 id,
